More rationalizing of Trading API and impl class locations
Converted MarketOrder and MarketOrderBook to interfaces and provided
default impls in Exchange Adapter project.
