+++ bxbot-exchanges/src/main/java/com/gazbert/bxbot/exchanges/BitfinexExchangeAdapter.java
@@ -26,6 +26,8 @@ package com.gazbert.bxbot.exchanges;
 import com.gazbert.bxbot.exchange.api.AuthenticationConfig;
 import com.gazbert.bxbot.exchange.api.ExchangeAdapter;
 import com.gazbert.bxbot.exchange.api.ExchangeConfig;
+import com.gazbert.bxbot.exchanges.trading.api.impl.MarketOrderBookImpl;
+import com.gazbert.bxbot.exchanges.trading.api.impl.MarketOrderImpl;
 import com.gazbert.bxbot.trading.api.*;
 import com.google.common.base.MoreObjects;
 import com.google.gson.Gson;
@@ -182,7 +184,7 @@ public final class BitfinexExchangeAdapter extends AbstractExchangeAdapter imple
 
             final List<MarketOrder> buyOrders = new ArrayList<>();
             for (BitfinexMarketOrder bitfinexBuyOrder : orderBook.bids) {
-                final MarketOrder buyOrder = new MarketOrder(
+                final MarketOrder buyOrder = new MarketOrderImpl(
                         OrderType.BUY,
                         bitfinexBuyOrder.price,
                         bitfinexBuyOrder.amount,
@@ -192,7 +194,7 @@ public final class BitfinexExchangeAdapter extends AbstractExchangeAdapter imple
 
             final List<MarketOrder> sellOrders = new ArrayList<>();
             for (BitfinexMarketOrder bitfinexSellOrder : orderBook.asks) {
-                final MarketOrder sellOrder = new MarketOrder(
+                final MarketOrder sellOrder = new MarketOrderImpl(
                         OrderType.SELL,
                         bitfinexSellOrder.price,
                         bitfinexSellOrder.amount,
@@ -200,7 +202,7 @@ public final class BitfinexExchangeAdapter extends AbstractExchangeAdapter imple
                 sellOrders.add(sellOrder);
             }
 
-            return new MarketOrderBook(marketId, sellOrders, buyOrders);
+            return new MarketOrderBookImpl(marketId, sellOrders, buyOrders);
 
         } catch (ExchangeNetworkException | TradingApiException e) {
             throw e;
