+++ bxbot-exchanges/src/main/java/com/gazbert/bxbot/exchanges/HuobiExchangeAdapter.java
@@ -27,6 +27,8 @@ import com.gazbert.bxbot.exchange.api.AuthenticationConfig;
 import com.gazbert.bxbot.exchange.api.ExchangeAdapter;
 import com.gazbert.bxbot.exchange.api.ExchangeConfig;
 import com.gazbert.bxbot.exchange.api.OptionalConfig;
+import com.gazbert.bxbot.exchanges.trading.api.impl.MarketOrderBookImpl;
+import com.gazbert.bxbot.exchanges.trading.api.impl.MarketOrderImpl;
 import com.gazbert.bxbot.trading.api.*;
 import com.google.common.base.MoreObjects;
 import com.google.gson.*;
@@ -436,7 +438,7 @@ public final class HuobiExchangeAdapter extends AbstractExchangeAdapter implemen
             // adapt BUYs
             final List<MarketOrder> buyOrders = new ArrayList<>();
             for (HuobiMarketOrder okCoinBuyOrder : orderBook.buys) {
-                final MarketOrder buyOrder = new MarketOrder(
+                final MarketOrder buyOrder = new MarketOrderImpl(
                         OrderType.BUY,
                         okCoinBuyOrder.price,
                         okCoinBuyOrder.amount,
@@ -447,7 +449,7 @@ public final class HuobiExchangeAdapter extends AbstractExchangeAdapter implemen
             // adapt SELLs
             final List<MarketOrder> sellOrders = new ArrayList<>();
             for (HuobiMarketOrder okCoinSellOrder : orderBook.sells) {
-                final MarketOrder sellOrder = new MarketOrder(
+                final MarketOrder sellOrder = new MarketOrderImpl(
                         OrderType.SELL,
                         okCoinSellOrder.price,
                         okCoinSellOrder.amount,
@@ -455,7 +457,7 @@ public final class HuobiExchangeAdapter extends AbstractExchangeAdapter implemen
                 sellOrders.add(sellOrder);
             }
 
-            return new MarketOrderBook(marketId, sellOrders, buyOrders);
+            return new MarketOrderBookImpl(marketId, sellOrders, buyOrders);
 
         } catch (ExchangeNetworkException | TradingApiException e) {
             throw e;
