+++ bxbot-exchanges/src/main/java/com/gazbert/bxbot/exchanges/KrakenExchangeAdapter.java
@@ -27,6 +27,8 @@ import com.gazbert.bxbot.exchange.api.AuthenticationConfig;
 import com.gazbert.bxbot.exchange.api.ExchangeAdapter;
 import com.gazbert.bxbot.exchange.api.ExchangeConfig;
 import com.gazbert.bxbot.exchange.api.OptionalConfig;
+import com.gazbert.bxbot.exchanges.trading.api.impl.MarketOrderBookImpl;
+import com.gazbert.bxbot.exchanges.trading.api.impl.MarketOrderImpl;
 import com.gazbert.bxbot.trading.api.*;
 import com.google.common.base.MoreObjects;
 import com.google.gson.*;
@@ -306,7 +308,7 @@ public final class KrakenExchangeAdapter extends AbstractExchangeAdapter impleme
 
                     final List<MarketOrder> buyOrders = new ArrayList<>();
                     for (KrakenMarketOrder krakenBuyOrder : krakenOrderBook.bids) {
-                        final MarketOrder buyOrder = new MarketOrder(
+                        final MarketOrder buyOrder = new MarketOrderImpl(
                                 OrderType.BUY,
                                 krakenBuyOrder.get(0),
                                 krakenBuyOrder.get(1),
@@ -316,7 +318,7 @@ public final class KrakenExchangeAdapter extends AbstractExchangeAdapter impleme
 
                     final List<MarketOrder> sellOrders = new ArrayList<>();
                     for (KrakenMarketOrder krakenSellOrder : krakenOrderBook.asks) {
-                        final MarketOrder sellOrder = new MarketOrder(
+                        final MarketOrder sellOrder = new MarketOrderImpl(
                                 OrderType.SELL,
                                 krakenSellOrder.get(0),
                                 krakenSellOrder.get(1),
@@ -324,7 +326,7 @@ public final class KrakenExchangeAdapter extends AbstractExchangeAdapter impleme
                         sellOrders.add(sellOrder);
                     }
 
-                    return new MarketOrderBook(marketId, sellOrders, buyOrders);
+                    return new MarketOrderBookImpl(marketId, sellOrders, buyOrders);
 
                 } else {
