+++ bxbot-exchanges/src/main/java/com/gazbert/bxbot/exchanges/OkCoinExchangeAdapter.java
@@ -27,6 +27,8 @@ import com.gazbert.bxbot.exchange.api.AuthenticationConfig;
 import com.gazbert.bxbot.exchange.api.ExchangeAdapter;
 import com.gazbert.bxbot.exchange.api.ExchangeConfig;
 import com.gazbert.bxbot.exchange.api.OptionalConfig;
+import com.gazbert.bxbot.exchanges.trading.api.impl.MarketOrderBookImpl;
+import com.gazbert.bxbot.exchanges.trading.api.impl.MarketOrderImpl;
 import com.gazbert.bxbot.trading.api.*;
 import com.google.common.base.MoreObjects;
 import com.google.gson.Gson;
@@ -334,7 +336,7 @@ public final class OkCoinExchangeAdapter extends AbstractExchangeAdapter impleme
 
             final List<MarketOrder> buyOrders = new ArrayList<>();
             for (OKCoinMarketOrder okCoinBuyOrder : orderBook.bids) {
-                final MarketOrder buyOrder = new MarketOrder(
+                final MarketOrder buyOrder = new MarketOrderImpl(
                         OrderType.BUY,
                         okCoinBuyOrder.get(0),
                         okCoinBuyOrder.get(1),
@@ -344,7 +346,7 @@ public final class OkCoinExchangeAdapter extends AbstractExchangeAdapter impleme
 
             final List<MarketOrder> sellOrders = new ArrayList<>();
             for (OKCoinMarketOrder okCoinSellOrder : orderBook.asks) {
-                final MarketOrder sellOrder = new MarketOrder(
+                final MarketOrder sellOrder = new MarketOrderImpl(
                         OrderType.SELL,
                         okCoinSellOrder.get(0),
                         okCoinSellOrder.get(1),
@@ -364,7 +366,7 @@ public final class OkCoinExchangeAdapter extends AbstractExchangeAdapter impleme
                 }
             });
 
-            return new MarketOrderBook(marketId, sellOrders, buyOrders);
+            return new MarketOrderBookImpl(marketId, sellOrders, buyOrders);
 
         } catch (ExchangeNetworkException | TradingApiException e) {
             throw e;
