+++ bxbot-exchanges/src/main/java/com/gazbert/bxbot/exchanges/TestExchangeAdapter.java
@@ -25,6 +25,8 @@ package com.gazbert.bxbot.exchanges;
 
 import com.gazbert.bxbot.exchange.api.ExchangeAdapter;
 import com.gazbert.bxbot.exchange.api.ExchangeConfig;
+import com.gazbert.bxbot.exchanges.trading.api.impl.MarketOrderBookImpl;
+import com.gazbert.bxbot.exchanges.trading.api.impl.MarketOrderImpl;
 import com.gazbert.bxbot.trading.api.*;
 import com.google.common.base.MoreObjects;
 import com.google.gson.*;
@@ -99,7 +101,7 @@ public final class TestExchangeAdapter extends AbstractExchangeAdapter implement
             final List<MarketOrder> buyOrders = new ArrayList<>();
             final List<List<BigDecimal>> bitstampBuyOrders = bitstampOrderBook.bids;
             for (final List<BigDecimal> order : bitstampBuyOrders) {
-                final MarketOrder buyOrder = new MarketOrder(
+                final MarketOrder buyOrder = new MarketOrderImpl(
                         OrderType.BUY,
                         order.get(0), // price
                         order.get(1), // quantity
@@ -110,7 +112,7 @@ public final class TestExchangeAdapter extends AbstractExchangeAdapter implement
             final List<MarketOrder> sellOrders = new ArrayList<>();
             final List<List<BigDecimal>> bitstampSellOrders = bitstampOrderBook.asks;
             for (final List<BigDecimal> order : bitstampSellOrders) {
-                final MarketOrder sellOrder = new MarketOrder(
+                final MarketOrder sellOrder = new MarketOrderImpl(
                         OrderType.SELL,
                         order.get(0), // price
                         order.get(1), // quantity
@@ -118,7 +120,7 @@ public final class TestExchangeAdapter extends AbstractExchangeAdapter implement
                 sellOrders.add(sellOrder);
             }
 
-            return new MarketOrderBook(marketId, sellOrders, buyOrders);
+            return new MarketOrderBookImpl(marketId, sellOrders, buyOrders);
 
         } catch (ExchangeNetworkException | TradingApiException e) {
             throw e;
