+++ bxbot-trading-api/src/main/java/com/gazbert/bxbot/trading/api/MarketOrder.java
@@ -23,8 +23,6 @@
 
 package com.gazbert.bxbot.trading.api;
 
-import com.google.common.base.MoreObjects;
-
 import java.math.BigDecimal;
 
 /**
@@ -38,123 +36,33 @@ import java.math.BigDecimal;
  * @author gazbert
  * @since 1.0
  */
-public final class MarketOrder {
-
-    /**
-     * The type of order.
-     * Value will be {@link OrderType#BUY} or {@link OrderType#SELL}.
-     */
-    private OrderType type;
-
-    /**
-     * The price of the order. This is usually in BTC or USD.
-     */
-    private BigDecimal price;
-
-    /**
-     * The quantity of the order. This is usually the amount of the other currency you want to trade for BTC/USD.
-     */
-    private BigDecimal quantity;
-
-    /**
-     * The total value of the order (price * quantity). This is usually in BTC or USD.
-     */
-    private BigDecimal total;
-
-    /**
-     * Constructor builds a Market Order.
-     *
-     * @param type     Type of order. Value must be {@link OrderType#BUY} or {@link OrderType#SELL}.
-     * @param price    Price of the order. This is usually in BTC or USD.
-     * @param quantity Quantity of the order. This is usually the amount of the other currency you want to trade for BTC/USD.
-     * @param total    Total value of order (price * quantity). This is usually in BTC or USD.
-     */
-    public MarketOrder(OrderType type, BigDecimal price, BigDecimal quantity, BigDecimal total) {
-        this.type = type;
-        this.price = price;
-        this.quantity = quantity;
-        this.total = total;
-    }
+public interface MarketOrder {
 
     /**
      * Returns the type of order. Value will be {@link OrderType#BUY} or {@link OrderType#SELL}.
      *
      * @return the type of order.
      */
-    public OrderType getType() {
-        return type;
-    }
-
-    /**
-     * Sets the type of order. Value must be {@link OrderType#BUY} or {@link OrderType#SELL}.
-     *
-     * @param type the order type.
-     */
-    public void setType(OrderType type) {
-        this.type = type;
-    }
+    OrderType getType();
 
     /**
      * Returns the price of the order. This is usually in BTC or USD.
      *
      * @return Price of the order.
      */
-    public BigDecimal getPrice() {
-        return price;
-    }
-
-    /**
-     * Sets the price of the order. This is usually in BTC or USD.
-     *
-     * @param price the price of the order.
-     */
-    public void setPrice(BigDecimal price) {
-        this.price = price;
-    }
+    BigDecimal getPrice();
 
     /**
      * Returns the quantity of the order. This is usually the amount of the other currency you want to trade for BTC/USD.
      *
      * @return Quantity of the order.
      */
-    public BigDecimal getQuantity() {
-        return quantity;
-    }
-
-    /**
-     * Sets the quantity of the order. This is usually the amount of the other currency you want to trade for BTC/USD.
-     *
-     * @param quantity the quantity of the order.
-     */
-    public void setQuantity(BigDecimal quantity) {
-        this.quantity = quantity;
-    }
+    BigDecimal getQuantity();
 
     /**
      * Returns the total value of order (price * quantity). This is usually in BTC or USD.
      *
      * @return Total value of order (price * quantity).
      */
-    public BigDecimal getTotal() {
-        return total;
-    }
-
-    /**
-     * Sets the total value of order (price * quantity). This is usually in BTC or USD.
-     *
-     * @param total the total value of the order.
-     */
-    public void setTotal(BigDecimal total) {
-        this.total = total;
-    }
-
-    @Override
-    public String toString() {
-        return MoreObjects.toStringHelper(this)
-                .add("type", type)
-                .add("price", price)
-                .add("quantity", quantity)
-                .add("total", total)
-                .toString();
-    }
+    BigDecimal getTotal();
 }
